Fixed Capital Gains Functionality
Removed lines of code that were causing errors. Not sure if I was using
a different excel file originally, but did not need to grab column names
from a lower row, they were correct from the start. Also data type was
correct from the start. Added directions to specify the files needed
since the 360 report is supposed to be used.